refactor: Add a Mission component for modularity
