+++ src/components/Missions.js
@@ -1,60 +1,52 @@
 import React, { useEffect } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
-import {
-  Table, Container, Badge, Button,
-} from 'react-bootstrap';
-import {
-  fetchMissions,
-  toggleMission,
-} from '../redux/missions/missionActionCreator';
-import { missions } from '../redux/missions/mission';
+import { useDispatch } from 'react-redux';
+import { Table, Container } from 'react-bootstrap';
+import { fetchMissions } from '../redux/missions/missionActionCreator';
+// import { missions } from '../redux/missions/mission';
+import Mission from './Mission';
 
 const Missions = () => {
   const dispatch = useDispatch();
-  const allMissions = useSelector(missions);
+  // const allMissions = useSelector(missions);
 
   useEffect(() => {
     dispatch(fetchMissions);
   }, [fetchMissions]);
 
-  const joinOrLeaveMission = (e) => {
-    dispatch(toggleMission({ mission_id: e.target.id }));
-  };
-
-  const missionComponents = allMissions.map((mission) => (
-    <tr key={mission.mission_id}>
-      <td>
-        <b>{mission.mission_name}</b>
-      </td>
-      <td>{mission.description}</td>
-      <td className="px-4 align-middle">
-        {mission.reserved ? (
-          <Badge className="bg-info">Active Member</Badge>
-        ) : (
-          <Badge className="bg-secondary">NOT A MEMBER</Badge>
-        )}
-      </td>
-      <td className="px-4 align-middle">
-        {mission.reserved ? (
-          <Button
-            variant="outline-danger"
-            id={mission.mission_id}
-            onClick={joinOrLeaveMission}
-          >
-            Leave&nbsp;Mission
-          </Button>
-        ) : (
-          <Button
-            variant="outline-secondary"
-            id={mission.mission_id}
-            onClick={joinOrLeaveMission}
-          >
-            Join&nbsp;Mission
-          </Button>
-        )}
-      </td>
-    </tr>
-  ));
+  // const missionComponents = allMissions.map((mission) => (
+  //   <tr key={mission.mission_id}>
+  //     <td>
+  //       <b>{mission.mission_name}</b>
+  //     </td>
+  //     <td>{mission.description}</td>
+  //     <td className="px-4 align-middle">
+  //       {mission.reserved ? (
+  //         <Badge className="bg-info">Active Member</Badge>
+  //       ) : (
+  //         <Badge className="bg-secondary">NOT A MEMBER</Badge>
+  //       )}
+  //     </td>
+  //     <td className="px-4 align-middle">
+  //       {mission.reserved ? (
+  //         <Button
+  //           variant="outline-danger"
+  //           id={mission.mission_id}
+  //           onClick={joinOrLeaveMission}
+  //         >
+  //           Leave&nbsp;Mission
+  //         </Button>
+  //       ) : (
+  //         <Button
+  //           variant="outline-secondary"
+  //           id={mission.mission_id}
+  //           onClick={joinOrLeaveMission}
+  //         >
+  //           Join&nbsp;Mission
+  //         </Button>
+  //       )}
+  //     </td>
+  //   </tr>
+  // ));
   return (
     <Container fluid className="table-responsive-sm">
       <Table className="my-3 table-bordered table-striped">
@@ -72,7 +64,9 @@ const Missions = () => {
             <td />
           </tr>
         </thead>
-        <tbody>{missionComponents}</tbody>
+        <tbody>
+          <Mission />
+        </tbody>
       </Table>
     </Container>
   );
