+++ src/redux/missions/missionActionCreator.js
@@ -10,14 +10,12 @@ export const fetchMissions = async (dispatch) => {
   const fetchMissions = await fetch(`${missionBaseURL}`);
   const returnList = await fetchMissions.json();
   const missions = [];
-  returnList.map((mission) =>
-    missions.push({
-      mission_id: mission.mission_id,
-      mission_name: mission.mission_name,
-      description: mission.description,
-      wikipedia: mission.wikipedia,
-    })
-  );
+  returnList.map((mission) => missions.push({
+    mission_id: mission.mission_id,
+    mission_name: mission.mission_name,
+    description: mission.description,
+    wikipedia: mission.wikipedia,
+  }));
   dispatch({
     type: actions.ADD_MISSION,
     payload: missions,
